Match (admin + client side)
- Admin: add Statistic function
- Client:
   + List Matches
   + View one Match

+++ FootballFlick/Areas/Admin/Controllers/ClubPlayerController.cs
@@ -24,7 +24,7 @@ namespace FootballFlick.Areas.Admin.Controllers
 
         //Delete an ClubPlayer
         [HttpDelete]
-        public ActionResult Delete(long clubId, long playerId)
+        public ActionResult Delete(long clubId, int playerId)
         {
             new ClubPlayerDao().Delete(clubId, playerId);
             return RedirectToAction("Index");
